requirements.txt
Flask==0.12.2
psycopg2==2.7.3.1
PyJWT==1.5.3
